+++ src/parser/block/paragraph/mention.ts
@@ -1,13 +1,13 @@
 import { ParagraphParser } from '../../block';
 import { subsequence, some, block, firstline, bind } from '../../../combinator';
 import { address } from './mention/address';
-import { quote } from './mention/quote';
+import { quotation } from './mention/quotation';
 import { html } from 'typed-dom';
 
 export const mention: ParagraphParser.MentionParser = block(bind(
   subsequence([
     some(address),
-    quote,
+    quotation,
   ]),
   (ns, rest) => [
     ns.reduceRight((acc, node) => {